+++ src/features/extension/info/components/SourceCard.tsx
@@ -24,13 +24,23 @@ import { StyledGroupItemWrapper } from '@/base/components/virtuoso/StyledGroupIt
 import { getErrorMessage } from '@/lib/HelperFunctions.ts';
 import { makeToast } from '@/base/utils/Toast.ts';
 import { createUpdateSourceMetadata, useGetSourceMetadata } from '@/features/source/services/SourceMetadata.ts';
-import type { SourceConfigurableInfo, SourceIdInfo, SourceLanguageInfo } from '@/features/source/Source.types.ts';
+import type {
+    SourceConfigurableInfo,
+    SourceIdInfo,
+    SourceLanguageInfo,
+    SourceMetaInfo,
+} from '@/features/source/Source.types.ts';
+import { useMemo } from 'react';
 
-export const SourceCard = (source: SourceIdInfo & SourceLanguageInfo & SourceConfigurableInfo) => {
+export const SourceCard = (source: SourceIdInfo & SourceMetaInfo & SourceLanguageInfo & SourceConfigurableInfo) => {
     const { id, isConfigurable } = source;
 
     const { t } = useLingui();
     const { isEnabled } = useGetSourceMetadata(source);
+    const { languages, setLanguages } = Sources.useLanguages();
+
+    const isLanguageEnabled = useMemo(() => Sources.isLanguageEnabled(source, languages), [languages, source.lang]);
+    const finalIsEnabled = useMemo(() => Sources.isEnabled(source, languages), [source, languages]);
 
     const updateSetting = createUpdateSourceMetadata(source, (e) =>
         makeToast(t`Failed to save changes`, 'error', getErrorMessage(e)),
@@ -39,7 +49,21 @@ export const SourceCard = (source: SourceIdInfo & SourceLanguageInfo & SourceCon
     return (
         <StyledGroupItemWrapper key={id} sx={{ px: 0 }}>
             <Card>
-                <CardActionArea onClick={() => updateSetting('isEnabled', !isEnabled)}>
+                <CardActionArea
+                    onClick={() => {
+                        if (!isLanguageEnabled) {
+                            setLanguages([...languages, source.lang]);
+
+                            if (!isEnabled) {
+                                updateSetting('isEnabled', !isEnabled);
+                            }
+
+                            return;
+                        }
+
+                        updateSetting('isEnabled', !isEnabled);
+                    }}
+                >
                     <ListCardContent>
                         <Typography variant="h6" component="h3" sx={{ flexGrow: 1 }}>
                             {translateExtensionLanguage(Sources.getLanguage(source))}
@@ -57,7 +81,7 @@ export const SourceCard = (source: SourceIdInfo & SourceLanguageInfo & SourceCon
                                 </IconButton>
                             </CustomTooltip>
                         )}
-                        <Switch checked={isEnabled} />
+                        <Switch checked={finalIsEnabled} />
                     </ListCardContent>
                 </CardActionArea>
             </Card>